api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-cosmos
azure-identity
cryptography==43.0.3
openai
pydantic
azure-keyvault-secrets
azure-storage-blob
